Redid move of plugin-api to the plugin-api module. Missed in the first rebase merge :(
