+++ src/main/groovy/se/diabol/scrolls/plugins/GitHubPlugin.groovy
@@ -1,7 +1,6 @@
 package se.diabol.scrolls.plugins
 
 import com.mashape.unirest.http.Unirest
-import se.diabol.scrolls.engine.ScrollsPlugin
 
 class GitHubPlugin implements ScrollsPlugin {
     def config
