+++ src/main/groovy/se/diabol/scrolls/plugins/JiraPlugin.groovy
@@ -1,7 +1,6 @@
 package se.diabol.scrolls.plugins
 
 import groovyx.net.http.*
-import se.diabol.scrolls.engine.ScrollsPlugin
 
 import java.nio.charset.StandardCharsets
